requirement.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
defusedxml==0.7.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.3
django-allauth==0.51.0
django-gravatar2==1.4.4
django-heroku==0.3.1
django-login-required-middleware==0.8.0
django-shortuuidfield==0.1.3
gunicorn==20.1.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==2.5.2
psycopg2==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==3.3.3
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
twilio==7.15.4
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.3.0
wrapt==1.14.1
